Use tox in Makefile targets
Modify the Makefile to point at the appropriate tox targets
so that tox and Make output can be equivalent. This involves
mapping the lint target to the pep8 target and the test target
to the py27 target.

Update tox.ini to not use site-packages as the PyYAML installed
version conflicts with pip requirements.

Change-Id: I32c2829af2a44e25c429bb603098a1249d415356